Move doSearch function to util
