+++ lib/backend.js
@@ -1,11 +1,10 @@
 const express = require('express')
-const _ = require('lodash')
 const {promisify} = require('es6-promisify')
 const path = require('path')
 const fs = require('fs')
 const os = require('os')
 const errorhandler = require('errorhandler')
-const { extractKey, extractKeysForItems, parseKey } = require('./util')
+const { extractKey, extractKeysForItems, parseKey, doSearch } = require('./util')
 const bodyParser = require('body-parser')
 const pickBy = require('lodash/pickBy')
 const omit = require('lodash/omit')
@@ -269,66 +268,8 @@ exports.createServer = (dynamodb, docClient) => {
     })
   })
 
-  const doSearch = (docClient, tableName, scanParams, limit, startKey, done, progress,
-                    readOperation = 'scan') => {
-    limit = typeof limit !== 'undefined' ? limit : null
-    startKey = typeof startKey !== 'undefined' ? startKey : null
-    let params = { TableName: tableName }
-    if (typeof scanParams !== 'undefined' && scanParams) {
-      params = _.assign(params, scanParams)
-    }
-    if (limit != null) {params.Limit = limit}
-    if (startKey != null) {params.ExclusiveStartKey = startKey}
-    const items = []
-    const processNextBite = function(err, items, nextKey) {
-      if (!err && nextKey) {
-        params.ExclusiveStartKey = nextKey
-        getNextBite(params, items, processNextBite)
-      } else {
-        if (done) {done(err, items)}
-      }
-    }
-
-    const readMethod = {
-      scan: docClient.scan,
-      query: docClient.query
-    }[readOperation].bind(docClient)
-
-    const getNextBite = function(params, items, callback) {
-      readMethod(params, function(err, data) {
-        if (err !== null) {
-          callback(err, items, null)
-          return
-        }
-
-        if (data && data.Items && data.Items.length > 0) {
-          items = items.concat(data.Items)
-        }
-
-        let lastStartKey = null
-        if (data) {
-          lastStartKey = data.LastEvaluatedKey
-        }
-
-        if (progress) {
-          const stop = progress(err, data.Items, lastStartKey)
-          if (!stop) {
-            callback(err, items, lastStartKey)
-          } else {
-            if (done) {
-              done(err, items)
-            }
-          }
-        } else {
-          callback(err, items, lastStartKey)
-        }
-      })
-    }
-    getNextBite(params, items, processNextBite)
-  }
-
-  const getPage = (docClient, keySchema, TableName, scanParams, pageSize, startKey,
-                   done) => {
+  const getPage = (docClient, keySchema, TableName, scanParams, pageSize,
+                   startKey, done) => {
     let pageItems = []
     doSearch(
       docClient,
@@ -338,24 +279,21 @@ exports.createServer = (dynamodb, docClient) => {
       startKey,
       err => {
         let nextKey = null
-        if (_.size(pageItems) > pageSize) {
+        if (pageItems.length > pageSize) {
           pageItems = pageItems.slice(0, pageSize)
           nextKey = extractKey(pageItems[pageSize - 1], keySchema)
         }
         done(pageItems, err, nextKey)
       },
       function(err, items, lastStartKey) {
-        for (
-          let i = 0;
-          i < items.length && _.size(pageItems) < pageSize + 1;
-          i++
-        ) {
+        for (let i = 0; i < items.length && pageItems.length < pageSize + 1; i++) {
           pageItems.push(items[i])
         }
+
         // If there is more items to query (!lastStartKey) then don't stop until
         // we are over pageSize count. Stopping at exactly pageSize count would
         // not extract key of last item later and make pagination not work.
-        return _.size(pageItems) > pageSize || !lastStartKey
+        return pageItems.length > pageSize || !lastStartKey
       }
     )
   }
